Use PodSpec instead of JobSpec
This commit changes the way the admission handler works. Instead of
checking if the object is a 'Job', we now only work with 'Pods'. The
reasoning for this is that a pod will simplify how this works and enable
this controller to work with more than just one type of object.

Aside from this, annotations have been removed in favour of just one
label.

Signed-off-by: Matei David <matei@buoyant.io>